restructure existing routes and endpoints for easy tracking for succeeding related endpoints
