--- srcs/database/routes/avatars/schemas/avatar.json
@@ -1,15 +1,0 @@
-{
-  "type": "object",
-  "$id": "schema:avatar:upload",
-  "required": [
-    "userId",
-    "avatar"
-  ],
-  "properties": {
-    "userId": { "type": "integer" },
-    "avatar": {
-      "type": "string",
-      "format": "binary"
-    }
-  }
-}
