--- srcs/database/routes/me/schemas/userProfile.json
@@ -1,25 +1,0 @@
-{
-  "type": "object",
-  "$id": "schema:me:userProfile",
-  "required": [
-    "username",
-    "email",
-    "nickname",
-    "created",
-    "avatar"
-  ],
-  "properties": {
-    "username": { "type": "string" },
-    "email": { "type": "string", "format": "email" },
-    "nickname": { "type": "string" },
-    "created": { "type": "string", "format": "date-time" },
-    "avatar": {
-      "type": "object",
-      "$id": "#/properties/avatar",
-      "properties": {
-        "url": { "type": "string", "format": "uri" }
-      },
-      "required": ["url"]
-    }
-  }
-}
