--- srcs/database/routes/user/schemas/getProfile.json
@@ -1,8 +1,0 @@
-{
-  "type": "object",
-  "$id": "schema:user:getProfile",
-  "required": [ "id" ],
-  "properties": {
-    "id": { "type": "string" }
-  }
-}
